requirements.txt
psycopg2==2.9.9
PyYAML==6.0.1
sqlparse==0.5.0
alembic==1.13.1
SQLAlchemy==2.0.30
overpy==0.7
python-telegram-bot==21.0.1
python-dotenv==1.0.1
geopy==2.4.1
